+++ training/dataset/dim3/dataset_acdc.py
@@ -76,8 +76,8 @@ class CMRDataset(Dataset):
 
     def preprocess(self, itk_img, itk_lab):
         
-        img = sitk.GetArrayFromImage(itk_img)
-        lab = sitk.GetArrayFromImage(itk_lab)
+        img = sitk.GetArrayFromImage(itk_img).astype(np.float32)
+        lab = sitk.GetArrayFromImage(itk_lab).astype(np.uint8)
 
         max98 = np.percentile(img, 98)
         img = np.clip(img, 0, max98)
@@ -152,7 +152,7 @@ class CMRDataset(Dataset):
         assert tensor_img.shape == tensor_lab.shape
 
         if self.mode == 'train':
-            return tensor_img, tensor_lab
+            return tensor_img, tensor_lab.to(torch.int8)
         else:
             return tensor_img, tensor_lab, np.array(self.spacing_list[idx])
